@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,7 +26,6 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser;
@@ -43,8 +43,6 @@ use Smalot\PdfParser\Element\ElementXRef;
 
 /**
  * Class Element
- *
- * @package Smalot\PdfParser
  */
 class Element
 {
@@ -53,47 +51,35 @@ class Element
      */
     protected $document = null;
 
-    /**
-     * @var mixed
-     */
     protected $value = null;
 
     /**
-     * @param mixed    $value
      * @param Document $document
      */
     public function __construct($value, Document $document = null)
     {
-        $this->value    = $value;
+        $this->value = $value;
         $this->document = $document;
     }
 
-    /**
-     *
-     */
     public function init()
     {
-
     }
 
     /**
-     * @param mixed $value
-     *
      * @return bool
      */
     public function equals($value)
     {
-        return ($value == $this->value);
+        return $value == $this->value;
     }
 
     /**
-     * @param mixed $value
-     *
      * @return bool
      */
     public function contains($value)
     {
-        if (is_array($this->value)) {
+        if (\is_array($this->value)) {
             /** @var Element $val */
             foreach ($this->value as $val) {
                 if ($val->equals($value)) {
@@ -102,14 +88,11 @@ class Element
             }
 
             return false;
-        } else {
-            return $this->equals($value);
         }
+
+        return $this->equals($value);
     }
 
-    /**
-     * @return mixed
-     */
     public function getContent()
     {
         return $this->value;
@@ -120,7 +103,7 @@ class Element
      */
     public function __toString()
     {
-        return (string)($this->value);
+        return (string) ($this->value);
     }
 
     /**
@@ -129,14 +112,15 @@ class Element
      * @param int      $position
      *
      * @return array
+     *
      * @throws \Exception
      */
     public static function parse($content, Document $document = null, &$position = 0)
     {
-        $args        = func_get_args();
+        $args = \func_get_args();
         $only_values = isset($args[3]) ? $args[3] : false;
-        $content     = trim($content);
-        $values      = array();
+        $content = trim($content);
+        $values = [];
 
         do {
             $old_position = $position;
@@ -145,12 +129,12 @@ class Element
                 if (!preg_match('/^\s*(?P<name>\/[A-Z0-9\._]+)(?P<value>.*)/si', substr($content, $position), $match)) {
                     break;
                 } else {
-                    $name     = ltrim($match['name'], '/');
-                    $value    = $match['value'];
-                    $position = strpos($content, $value, $position + strlen($match['name']));
+                    $name = ltrim($match['name'], '/');
+                    $value = $match['value'];
+                    $position = strpos($content, $value, $position + \strlen($match['name']));
                 }
             } else {
-                $name  = count($values);
+                $name = \count($values);
                 $value = substr($content, $position);
             }
 
@@ -178,7 +162,7 @@ class Element
                 $position = $old_position;
                 break;
             }
-        } while ($position < strlen($content));
+        } while ($position < \strlen($content));
 
         return $values;
     }
