@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,28 +26,24 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser\Element;
 
-use Smalot\PdfParser\Element;
 use Smalot\PdfParser\Document;
 
 /**
  * Class ElementDate
- *
- * @package Smalot\PdfParser\Element
  */
 class ElementDate extends ElementString
 {
     /**
      * @var array
      */
-    protected static $formats = array(
-        4  => 'Y',
-        6  => 'Ym',
-        8  => 'Ymd',
+    protected static $formats = [
+        4 => 'Y',
+        6 => 'Ym',
+        8 => 'Ymd',
         10 => 'YmdH',
         12 => 'YmdHi',
         14 => 'YmdHis',
@@ -54,7 +51,7 @@ class ElementDate extends ElementString
         17 => 'YmdHisO',
         18 => 'YmdHisO',
         19 => 'YmdHisO',
-    );
+    ];
 
     /**
      * @var string
@@ -63,15 +60,14 @@ class ElementDate extends ElementString
 
     /**
      * @param \DateTime $value
-     * @param Document  $document
      */
-    public function __construct($value, Document $document = null)
+    public function __construct($value)
     {
         if (!($value instanceof \DateTime)) {
             throw new \Exception('DateTime required.');
         }
 
-        parent::__construct($value, null);
+        parent::__construct($value);
     }
 
     /**
@@ -83,8 +79,6 @@ class ElementDate extends ElementString
     }
 
     /**
-     * @param mixed $value
-     *
      * @return bool
      */
     public function equals($value)
@@ -95,7 +89,7 @@ class ElementDate extends ElementString
             $timestamp = strtotime($value);
         }
 
-        return ($timestamp == $this->value->getTimeStamp());
+        return $timestamp == $this->value->getTimeStamp();
     }
 
     /**
@@ -103,7 +97,7 @@ class ElementDate extends ElementString
      */
     public function __toString()
     {
-        return (string)($this->value->format($this->format));
+        return (string) ($this->value->format($this->format));
     }
 
     /**
@@ -125,18 +119,18 @@ class ElementDate extends ElementString
             if (preg_match('/^\d{4}(\d{2}(\d{2}(\d{2}(\d{2}(\d{2}(Z(\d{2,4})?|[\+-]?\d{2}(\d{2})?)?)?)?)?)?)?$/', $name)) {
                 if ($pos = strpos($name, 'Z')) {
                     $name = substr($name, 0, $pos + 1);
-                } elseif (strlen($name) == 18 && preg_match('/[^\+-]0000$/', $name)) {
-                    $name = substr($name, 0, -4) . '+0000';
+                } elseif (18 == \strlen($name) && preg_match('/[^\+-]0000$/', $name)) {
+                    $name = substr($name, 0, -4).'+0000';
                 }
 
-                $format = self::$formats[strlen($name)];
-                $date   = \DateTime::createFromFormat($format, $name, new \DateTimeZone('UTC'));
+                $format = self::$formats[\strlen($name)];
+                $date = \DateTime::createFromFormat($format, $name, new \DateTimeZone('UTC'));
             } else {
                 // special cases
                 if (preg_match('/^\d{1,2}-\d{1,2}-\d{4},?\s+\d{2}:\d{2}:\d{2}[\+-]\d{4}$/', $name)) {
-                    $name   = str_replace(',', '', $name);
+                    $name = str_replace(',', '', $name);
                     $format = 'n-j-Y H:i:sO';
-                    $date   = \DateTime::createFromFormat($format, $name, new \DateTimeZone('UTC'));
+                    $date = \DateTime::createFromFormat($format, $name, new \DateTimeZone('UTC'));
                 }
             }
 
@@ -144,10 +138,9 @@ class ElementDate extends ElementString
                 return false;
             }
 
-            $offset += strpos($content, '(D:') + strlen($match['name']) + 4; // 1 for '(D:' and ')'
-            $element = new self($date, $document);
+            $offset += strpos($content, '(D:') + \strlen($match['name']) + 4; // 1 for '(D:' and ')'
 
-            return $element;
+            return new self($date);
         }
 
         return false;
