@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,21 +26,17 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser;
 
+use Smalot\PdfParser\Encoding\WinAnsiEncoding;
+
 /**
  * Class Font
- *
- * @package Smalot\PdfParser
  */
 class Font extends PDFObject
 {
-    /**
-     *
-     */
     const MISSING = '?';
 
     /**
@@ -52,9 +49,6 @@ class Font extends PDFObject
      */
     protected $tableSizes = null;
 
-    /**
-     *
-     */
     public function init()
     {
         // Load translate table.
@@ -66,7 +60,7 @@ class Font extends PDFObject
      */
     public function getName()
     {
-        return $this->has('BaseFont') ? (string)$this->get('BaseFont') : '[Unknown]';
+        return $this->has('BaseFont') ? (string) $this->get('BaseFont') : '[Unknown]';
     }
 
     /**
@@ -74,7 +68,7 @@ class Font extends PDFObject
      */
     public function getType()
     {
-        return (string)$this->header->get('Subtype');
+        return (string) $this->header->get('Subtype');
     }
 
     /**
@@ -82,11 +76,11 @@ class Font extends PDFObject
      */
     public function getDetails($deep = true)
     {
-        $details = array();
+        $details = [];
 
-        $details['Name']     = $this->getName();
-        $details['Type']     = $this->getType();
-        $details['Encoding'] = ($this->has('Encoding') ? (string)$this->get('Encoding') : 'Ansi');
+        $details['Name'] = $this->getName();
+        $details['Type'] = $this->getType();
+        $details['Encoding'] = ($this->has('Encoding') ? (string) $this->get('Encoding') : 'Ansi');
 
         $details += parent::getDetails($deep);
 
@@ -97,19 +91,27 @@ class Font extends PDFObject
      * @param string $char
      * @param bool   $use_default
      *
-     * @return string
+     * @return string|bool
      */
     public function translateChar($char, $use_default = true)
     {
         $dec = hexdec(bin2hex($char));
 
-        if (array_key_exists($dec, $this->table)) {
-            $char = $this->table[$dec];
-        } else {
-            $char = ($use_default ? self::MISSING : $char);
+        if (\array_key_exists($dec, $this->table)) {
+            return $this->table[$dec];
         }
 
-        return $char;
+        // fallback for decoding single-byte ANSI characters that are not in the lookup table
+        $fallbackDecoded = $char;
+        if (
+            \strlen($char) < 2
+            && $this->has('Encoding')
+            && WinAnsiEncoding::class === $this->get('Encoding')->__toString()
+        ) {
+            $fallbackDecoded = self::uchr($dec);
+        }
+
+        return $use_default ? self::MISSING : $fallbackDecoded;
     }
 
     /**
@@ -119,7 +121,9 @@ class Font extends PDFObject
      */
     public static function uchr($code)
     {
-        return html_entity_decode('&#' . ((int)$code) . ';', ENT_NOQUOTES, 'UTF-8');
+        // html_entity_decode() will not work with UTF-16 or UTF-32 char entities,
+        // therefore, we use mb_convert_encoding() instead
+        return mb_convert_encoding('&#'.((int) $code).';', 'UTF-8', 'HTML-ENTITIES');
     }
 
     /**
@@ -127,19 +131,19 @@ class Font extends PDFObject
      */
     public function loadTranslateTable()
     {
-        if (!is_null($this->table)) {
+        if (null !== $this->table) {
             return $this->table;
         }
 
-        $this->table      = array();
-        $this->tableSizes = array(
+        $this->table = [];
+        $this->tableSizes = [
             'from' => 1,
-            'to'   => 1,
-        );
+            'to' => 1,
+        ];
 
         if ($this->has('ToUnicode')) {
             $content = $this->get('ToUnicode')->getContent();
-            $matches = array();
+            $matches = [];
 
             // Support for multiple spacerange sections
             if (preg_match_all('/begincodespacerange(?P<sections>.*?)endcodespacerange/s', $content, $matches)) {
@@ -148,10 +152,10 @@ class Font extends PDFObject
 
                     preg_match_all($regexp, $section, $matches);
 
-                    $this->tableSizes = array(
-                        'from' => max(1, strlen(current($matches['from'])) / 2),
-                        'to'   => max(1, strlen(current($matches['to'])) / 2),
-                    );
+                    $this->tableSizes = [
+                        'from' => max(1, \strlen(current($matches['from'])) / 2),
+                        'to' => max(1, \strlen(current($matches['to'])) / 2),
+                    ];
 
                     break;
                 }
@@ -164,7 +168,7 @@ class Font extends PDFObject
 
                     preg_match_all($regexp, $section, $matches);
 
-                    $this->tableSizes['from'] = max(1, strlen(current($matches['from'])) / 2);
+                    $this->tableSizes['from'] = max(1, \strlen(current($matches['from'])) / 2);
 
                     foreach ($matches['from'] as $key => $from) {
                         $parts = preg_split(
@@ -173,7 +177,7 @@ class Font extends PDFObject
                             0,
                             PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE
                         );
-                        $text  = '';
+                        $text = '';
                         foreach ($parts as $part) {
                             $text .= self::uchr(hexdec($part));
                         }
@@ -192,10 +196,10 @@ class Font extends PDFObject
 
                     foreach ($matches['from'] as $key => $from) {
                         $char_from = hexdec($from);
-                        $char_to   = hexdec($matches['to'][$key]);
-                        $offset    = hexdec($matches['offset'][$key]);
+                        $char_to = hexdec($matches['to'][$key]);
+                        $offset = hexdec($matches['offset'][$key]);
 
-                        for ($char = $char_from; $char <= $char_to; $char++) {
+                        for ($char = $char_from; $char <= $char_to; ++$char) {
                             $this->table[$char] = self::uchr($char - $char_from + $offset);
                         }
                     }
@@ -208,7 +212,7 @@ class Font extends PDFObject
 
                     foreach ($matches['from'] as $key => $from) {
                         $char_from = hexdec($from);
-                        $strings   = array();
+                        $strings = [];
 
                         preg_match_all('/<(?P<string>[0-9A-F]+)> */is', $matches['strings'][$key], $strings);
 
@@ -219,7 +223,7 @@ class Font extends PDFObject
                                 0,
                                 PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE
                             );
-                            $text  = '';
+                            $text = '';
                             foreach ($parts as $part) {
                                 $text .= self::uchr(hexdec($part));
                             }
@@ -250,15 +254,17 @@ class Font extends PDFObject
     public static function decodeHexadecimal($hexa, $add_braces = false)
     {
         // Special shortcut for XML content.
-        if (stripos($hexa, '<?xml') !== false) {
+        if (false !== stripos($hexa, '<?xml')) {
             return $hexa;
         }
 
-        $text  = '';
+        $text = '';
         $parts = preg_split('/(<[a-f0-9]+>)/si', $hexa, -1, PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE);
 
         foreach ($parts as $part) {
-            if (preg_match('/^<.*>$/', $part) && stripos($part, '<?xml') === false) {
+            if (preg_match('/^<.*>$/s', $part) && false === stripos($part, '<?xml')) {
+                // strip line breaks
+                $part = preg_replace("/[\r\n]/", '', $part);
                 $part = trim($part, '<>');
                 if ($add_braces) {
                     $text .= '(';
@@ -286,11 +292,11 @@ class Font extends PDFObject
     public static function decodeOctal($text)
     {
         $parts = preg_split('/(\\\\\d{3})/s', $text, -1, PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE);
-        $text  = '';
+        $text = '';
 
         foreach ($parts as $part) {
             if (preg_match('/^\\\\\d{3}$/', $part)) {
-                $text .= chr(octdec(trim($part, '\\')));
+                $text .= \chr(octdec(trim($part, '\\')));
             } else {
                 $text .= $part;
             }
@@ -300,18 +306,18 @@ class Font extends PDFObject
     }
 
     /**
-     * @param $text
+     * @param string $text
      *
      * @return string
      */
     public static function decodeEntities($text)
     {
         $parts = preg_split('/(#\d{2})/s', $text, -1, PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE);
-        $text  = '';
+        $text = '';
 
         foreach ($parts as $part) {
             if (preg_match('/^#\d{2}$/', $part)) {
-                $text .= chr(hexdec(trim($part, '#')));
+                $text .= \chr(hexdec(trim($part, '#')));
             } else {
                 $text .= $part;
             }
@@ -330,8 +336,8 @@ class Font extends PDFObject
         if (preg_match('/^\xFE\xFF/i', $text)) {
             // Strip U+FEFF byte order marker.
             $decode = substr($text, 2);
-            $text   = '';
-            $length = strlen($decode);
+            $text = '';
+            $length = \strlen($decode);
 
             for ($i = 0; $i < $length; $i += 2) {
                 $text .= self::uchr(hexdec(bin2hex(substr($decode, $i, 2))));
@@ -343,10 +349,12 @@ class Font extends PDFObject
 
     /**
      * @return int
+     *
+     * @todo Deprecated, use $this->config->getFontSpaceLimit() instead.
      */
     protected function getFontSpaceLimit()
     {
-        return -50;
+        return $this->config->getFontSpaceLimit();
     }
 
     /**
@@ -356,27 +364,22 @@ class Font extends PDFObject
      */
     public function decodeText($commands)
     {
+        $text = '';
         $word_position = 0;
-        $words         = array();
-        $unicode       = false;
-        $font_space    = $this->getFontSpaceLimit();
+        $words = [];
+        $unicode = false;
+        $font_space = $this->getFontSpaceLimit();
 
         foreach ($commands as $command) {
             switch ($command[PDFObject::TYPE]) {
                 case 'n':
-                    if (floatval(trim($command[PDFObject::COMMAND])) < $font_space) {
-                        $word_position = count($words);
+                    if ((float) (trim($command[PDFObject::COMMAND])) < $font_space) {
+                        $word_position = \count($words);
                     }
-                    continue(2);
-
+                    continue 2;
                 case '<':
                     // Decode hexadecimal.
-                    $text = self::decodeHexadecimal('<' . $command[PDFObject::COMMAND] . '>');
-
-                    if (mb_check_encoding($text, "UTF-8")) {
-                        $unicode = true;
-                    }
-
+                    $text = self::decodeHexadecimal('<'.$command[PDFObject::COMMAND].'>');
                     break;
 
                 default:
@@ -386,8 +389,8 @@ class Font extends PDFObject
 
             // replace escaped chars
             $text = str_replace(
-                array('\\\\', '\(', '\)', '\n', '\r', '\t', '\ '),
-                array('\\', '(', ')', "\n", "\r", "\t", ' '),
+                ['\\\\', '\(', '\)', '\n', '\r', '\t', '\f', '\ '],
+                ['\\', '(', ')', "\n", "\r", "\t", "\f", ' '],
                 $text
             );
 
@@ -400,8 +403,7 @@ class Font extends PDFObject
         }
 
         foreach ($words as &$word) {
-            $loop_unicode = $unicode;
-            $word         = $this->decodeContent($word, $loop_unicode);
+            $word = $this->decodeContent($word);
         }
 
         return implode(' ', $words);
@@ -409,44 +411,42 @@ class Font extends PDFObject
 
     /**
      * @param string $text
-     * @param bool   $unicode
+     * @param bool   $unicode This parameter is deprecated and might be removed in a future release
      *
      * @return string
      */
-    protected function decodeContent($text, &$unicode)
+    public function decodeContent($text, &$unicode = null)
     {
         if ($this->has('ToUnicode')) {
-
             $bytes = $this->tableSizes['from'];
 
             if ($bytes) {
                 $result = '';
-                $length = strlen($text);
+                $length = \strlen($text);
 
                 for ($i = 0; $i < $length; $i += $bytes) {
                     $char = substr($text, $i, $bytes);
 
-                    if (($decoded = $this->translateChar($char, false)) !== false) {
+                    if (false !== ($decoded = $this->translateChar($char, false))) {
                         $char = $decoded;
                     } elseif ($this->has('DescendantFonts')) {
-
                         if ($this->get('DescendantFonts') instanceof PDFObject) {
-                            $fonts   = $this->get('DescendantFonts')->getHeader()->getElements();
+                            $fonts = $this->get('DescendantFonts')->getHeader()->getElements();
                         } else {
-                            $fonts   = $this->get('DescendantFonts')->getContent();
+                            $fonts = $this->get('DescendantFonts')->getContent();
                         }
                         $decoded = false;
 
                         foreach ($fonts as $font) {
-                            if ($font instanceof Font) {
-                                if (($decoded = $font->translateChar($char, false)) !== false) {
+                            if ($font instanceof self) {
+                                if (false !== ($decoded = $font->translateChar($char, false))) {
                                     $decoded = mb_convert_encoding($decoded, 'UTF-8', 'Windows-1252');
                                     break;
                                 }
                             }
                         }
 
-                        if ($decoded !== false) {
+                        if (false !== $decoded) {
                             $char = $decoded;
                         } else {
                             $char = mb_convert_encoding($char, 'UTF-8', 'Windows-1252');
@@ -459,22 +459,20 @@ class Font extends PDFObject
                 }
 
                 $text = $result;
-
-                // By definition, this code generates unicode chars.
-                $unicode = true;
             }
-        } elseif ($this->has('Encoding')) {
+        } elseif ($this->has('Encoding') && $this->get('Encoding') instanceof Encoding) {
             /** @var Encoding $encoding */
             $encoding = $this->get('Encoding');
-
-            if ($encoding instanceof Encoding) {
-                if ($unicode) {
-                    $chars  = preg_split(
-                        '//s' . ($unicode ? 'u' : ''),
+            $unicode = mb_check_encoding($text, 'UTF-8');
+            $result = '';
+            if ($unicode) {
+                $chars = preg_split(
+                        '//s'.($unicode ? 'u' : ''),
                         $text,
                         -1,
                         PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY
-                    );
+		);
+
                     $result = '';
                     
                     if (!$chars) {
@@ -506,7 +504,6 @@ class Font extends PDFObject
                     }
                 }
             }
-        }
 
         // Convert to unicode if not already done.
         if (!$unicode) {
@@ -515,9 +512,30 @@ class Font extends PDFObject
                 $this->get('Encoding')->equals('MacRomanEncoding')
             ) {
                 $text = iconv('macintosh', 'UTF-8', $text);
+
+                foreach ($chars as $char) {
+                    $dec_av = hexdec(bin2hex($char));
+                    $dec_ap = $encoding->translateChar($dec_av);
+                    $result .= self::uchr($dec_ap);
+                }
             } else {
-                $text = mb_convert_encoding($text, 'UTF-8', 'Windows-1252');
+                $length = \strlen($text);
+
+                for ($i = 0; $i < $length; ++$i) {
+                    $dec_av = hexdec(bin2hex($text[$i]));
+                    $dec_ap = $encoding->translateChar($dec_av);
+                    $result .= self::uchr($dec_ap);
+                }
             }
+            $text = $result;
+        } elseif ($this->get('Encoding') instanceof Element &&
+                  $this->get('Encoding')->equals('MacRomanEncoding')) {
+            // mb_convert_encoding does not support MacRoman/macintosh,
+            // so we use iconv() here
+            $text = iconv('macintosh', 'UTF-8', $text);
+        } elseif (!mb_check_encoding($text, 'UTF-8')) {
+            // don't double-encode strings already in UTF-8
+            $text = mb_convert_encoding($text, 'UTF-8', 'Windows-1252');
         }
 
         return $text;
