@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,7 +26,6 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser;
@@ -37,8 +37,6 @@ use Smalot\PdfParser\Element\ElementXRef;
 
 /**
  * Class Header
- *
- * @package Smalot\PdfParser
  */
 class Header
 {
@@ -53,19 +51,26 @@ class Header
     protected $elements = null;
 
     /**
-     * @param Element[] $elements   List of elements.
-     * @param Document  $document   Document.
+     * @param Element[] $elements list of elements
+     * @param Document  $document document
      */
-    public function __construct($elements = array(), Document $document = null)
+    public function __construct($elements = [], Document $document = null)
     {
         $this->elements = $elements;
         $this->document = $document;
     }
 
+    public function init()
+    {
+        foreach ($this->elements as $element) {
+            if ($element instanceof Element) {
+                $element->init();
+            }
+        }
+    }
+
     /**
      * Returns all elements.
-     *
-     * @return mixed
      */
     public function getElements()
     {
@@ -83,10 +88,10 @@ class Header
      */
     public function getElementTypes()
     {
-        $types = array();
+        $types = [];
 
         foreach ($this->elements as $key => $element) {
-            $types[$key] = get_class($element);
+            $types[$key] = \get_class($element);
         }
 
         return $types;
@@ -99,11 +104,11 @@ class Header
      */
     public function getDetails($deep = true)
     {
-        $values   = array();
+        $values = [];
         $elements = $this->getElements();
 
         foreach ($elements as $key => $element) {
-            if ($element instanceof Header && $deep) {
+            if ($element instanceof self && $deep) {
                 $values[$key] = $element->getDetails($deep);
             } elseif ($element instanceof PDFObject && $deep) {
                 $values[$key] = $element->getDetails(false);
@@ -128,11 +133,7 @@ class Header
      */
     public function has($name)
     {
-        if (array_key_exists($name, $this->elements)) {
-            return true;
-        } else {
-            return false;
-        }
+        return \array_key_exists($name, $this->elements);
     }
 
     /**
@@ -142,11 +143,11 @@ class Header
      */
     public function get($name)
     {
-        if (array_key_exists($name, $this->elements)) {
+        if (\array_key_exists($name, $this->elements)) {
             return $this->resolveXRef($name);
         }
 
-        return new ElementMissing(null, null);
+        return new ElementMissing();
     }
 
     /**
@@ -155,16 +156,17 @@ class Header
      * @param string $name
      *
      * @return Element|PDFObject
+     *
      * @throws \Exception
      */
     protected function resolveXRef($name)
     {
-        if (($obj = $this->elements[$name]) instanceof ElementXRef && !is_null($this->document)) {
+        if (($obj = $this->elements[$name]) instanceof ElementXRef && null !== $this->document) {
             /** @var ElementXRef $obj */
             $object = $this->document->getObjectById($obj->getId());
 
-            if (is_null($object)) {
-                return new ElementMissing(null, null);
+            if (null === $object) {
+                return new ElementMissing();
             }
 
             // Update elements list for future calls.
@@ -183,23 +185,23 @@ class Header
      */
     public static function parse($content, Document $document, &$position = 0)
     {
-        /** @var Header $header */
-        if (substr(trim($content), 0, 2) == '<<') {
+        /* @var Header $header */
+        if ('<<' == substr(trim($content), 0, 2)) {
             $header = ElementStruct::parse($content, $document, $position);
         } else {
             $elements = ElementArray::parse($content, $document, $position);
+            $header = new self([], $document);
+
             if ($elements) {
-                $header = new self($elements->getRawContent(), null);//$document);
-            } else {
-                $header = new self(array(), $document);
+                $header = new self($elements->getRawContent(), null);
             }
         }
 
         if ($header) {
             return $header;
-        } else {
-            // Build an empty header.
-            return new self(array(), $document);
         }
+
+        // Build an empty header.
+        return new self([], $document);
     }
 }
