@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,7 +26,6 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser;
@@ -35,8 +35,6 @@ use Smalot\PdfParser\XObject\Image;
 
 /**
  * Class PDFObject
- *
- * @package Smalot\PdfParser
  */
 class PDFObject
 {
@@ -51,7 +49,7 @@ class PDFObject
      *
      * @var array
      */
-    static $recursionStack = array();
+    public static $recursionStack = [];
 
     /**
      * @var Document
@@ -69,27 +67,33 @@ class PDFObject
     protected $content = null;
 
     /**
-     * @param Document $document
-     * @param Header   $header
-     * @param string   $content
+     * @var Config
      */
-    public function __construct(Document $document, Header $header = null, $content = null)
-    {
-        $this->document = $document;
-        $this->header   = !is_null($header) ? $header : new Header();
-        $this->content  = $content;
-    }
+    protected $config;
 
     /**
-     *
+     * @param Header $header
+     * @param string $content
+     * @param Config $config
      */
+    public function __construct(
+        Document $document,
+        Header $header = null,
+        $content = null,
+        Config $config = null
+    ) {
+        $this->document = $document;
+        $this->header = null !== $header ? $header : new Header();
+        $this->content = $content;
+        $this->config = $config;
+    }
+
     public function init()
     {
-
     }
 
     /**
-     * @return null|Header
+     * @return Header|null
      */
     public function getHeader()
     {
@@ -107,7 +111,7 @@ class PDFObject
     }
 
     /**
-     * @param $name
+     * @param string $name
      *
      * @return bool
      */
@@ -127,7 +131,7 @@ class PDFObject
     }
 
     /**
-     * @return null|string
+     * @return string|null
      */
     public function getContent()
     {
@@ -135,90 +139,90 @@ class PDFObject
     }
 
     /**
-     * @param $content
+     * @param string $content
      */
     public function cleanContent($content, $char = 'X')
     {
-        $char    = $char[0];
-        $content = str_replace(array('\\\\', '\\)', '\\('), $char . $char, $content);
+        $char = $char[0];
+        $content = str_replace(['\\\\', '\\)', '\\('], $char.$char, $content);
 
         // Remove image bloc with binary content
         preg_match_all('/\s(BI\s.*?(\sID\s).*?(\sEI))\s/s', $content, $matches, PREG_OFFSET_CAPTURE);
         foreach ($matches[0] as $part) {
-            $content = substr_replace($content, str_repeat($char, strlen($part[0])), $part[1], strlen($part[0]));
+            $content = substr_replace($content, str_repeat($char, \strlen($part[0])), $part[1], \strlen($part[0]));
         }
 
         // Clean content in square brackets [.....]
         preg_match_all('/\[((\(.*?\)|[0-9\.\-\s]*)*)\]/s', $content, $matches, PREG_OFFSET_CAPTURE);
         foreach ($matches[1] as $part) {
-            $content = substr_replace($content, str_repeat($char, strlen($part[0])), $part[1], strlen($part[0]));
+            $content = substr_replace($content, str_repeat($char, \strlen($part[0])), $part[1], \strlen($part[0]));
         }
 
         // Clean content in round brackets (.....)
         preg_match_all('/\((.*?)\)/s', $content, $matches, PREG_OFFSET_CAPTURE);
         foreach ($matches[1] as $part) {
-            $content = substr_replace($content, str_repeat($char, strlen($part[0])), $part[1], strlen($part[0]));
+            $content = substr_replace($content, str_repeat($char, \strlen($part[0])), $part[1], \strlen($part[0]));
         }
 
         // Clean structure
         if ($parts = preg_split('/(<|>)/s', $content, -1, PREG_SPLIT_NO_EMPTY | PREG_SPLIT_DELIM_CAPTURE)) {
             $content = '';
-            $level   = 0;
+            $level = 0;
             foreach ($parts as $part) {
-                if ($part == '<') {
-                    $level++;
+                if ('<' == $part) {
+                    ++$level;
                 }
 
-                $content .= ($level == 0 ? $part : str_repeat($char, strlen($part)));
+                $content .= (0 == $level ? $part : str_repeat($char, \strlen($part)));
 
-                if ($part == '>') {
-                    $level--;
+                if ('>' == $part) {
+                    --$level;
                 }
             }
         }
 
         // Clean BDC and EMC markup
         preg_match_all(
-            '/(\/[A-Za-z0-9\_]*\s*' . preg_quote($char) . '*BDC)/s',
+            '/(\/[A-Za-z0-9\_]*\s*'.preg_quote($char).'*BDC)/s',
             $content,
             $matches,
             PREG_OFFSET_CAPTURE
         );
         foreach ($matches[1] as $part) {
-            $content = substr_replace($content, str_repeat($char, strlen($part[0])), $part[1], strlen($part[0]));
+            $content = substr_replace($content, str_repeat($char, \strlen($part[0])), $part[1], \strlen($part[0]));
         }
 
         preg_match_all('/\s(EMC)\s/s', $content, $matches, PREG_OFFSET_CAPTURE);
         foreach ($matches[1] as $part) {
-            $content = substr_replace($content, str_repeat($char, strlen($part[0])), $part[1], strlen($part[0]));
+            $content = substr_replace($content, str_repeat($char, \strlen($part[0])), $part[1], \strlen($part[0]));
         }
 
         return $content;
     }
 
     /**
-     * @param $content
+     * @param string $content
      *
      * @return array
      */
     public function getSectionsText($content)
     {
-        $sections    = array();
-        $content     = ' ' . $content . ' ';
+        $sections = [];
+        $content = ' '.$content.' ';
         $textCleaned = $this->cleanContent($content, '_');
 
         // Extract text blocks.
         if (preg_match_all('/\s+BT[\s|\(|\[]+(.*?)\s*ET/s', $textCleaned, $matches, PREG_OFFSET_CAPTURE)) {
             foreach ($matches[1] as $part) {
-                $text    = $part[0];
-                if ($text === '') {
+                $text = $part[0];
+                if ('' === $text) {
                     continue;
                 }
-                $offset  = $part[1];
-                $section = substr($content, $offset, strlen($text));
+                $offset = $part[1];
+                $section = substr($content, $offset, \strlen($text));
 
                 // Removes BDC and EMC markup.
-                $section = preg_replace('/(\/[A-Za-z0-9]+\s*<<.*?)(>>\s*BDC)(.*?)(EMC\s+)/s', '${3}', $section . ' ');
+                $section = preg_replace('/(\/[A-Za-z0-9]+\s*<<.*?)(>>\s*BDC)(.*?)(EMC\s+)/s', '${3}', $section.' ');
 
                 $sections[] = $section;
             }
@@ -227,9 +231,9 @@ class PDFObject
         // Extract 'do' commands.
         if (preg_match_all('/(\/[A-Za-z0-9\.\-_]+\s+Do)\s/s', $textCleaned, $matches, PREG_OFFSET_CAPTURE)) {
             foreach ($matches[1] as $part) {
-                $text    = $part[0];
-                $offset  = $part[1];
-                $section = substr($content, $offset, strlen($text));
+                $text = $part[0];
+                $offset = $part[1];
+                $section = substr($content, $offset, \strlen($text));
 
                 $sections[] = $section;
             }
@@ -243,42 +247,46 @@ class PDFObject
 		$ret = $this->getText($page);
 		$this->cleanRecursionStack();
 		return $ret;
-	}
+    }
+
+    private function getDefaultFont(Page $page = null)
+    {
+        $fonts = [];
+        if (null !== $page) {
+            $fonts = $page->getFonts();
+        }
+
+        $fonts = array_merge($fonts, array_values($this->document->getFonts()));
+
+        if (\count($fonts) > 0) {
+            return reset($fonts);
+        }
+
+        return new Font($this->document);
+    }
 
     /**
-     * @param Page
+     * @param Page $page
      *
      * @return string
+     *
      * @throws \Exception
      */
     public function getText(Page $page = null)
     {
-        $text                = '';
-        $sections            = $this->getSectionsText($this->content);
-        $current_font = null;
+        $text = '';
+        $sections = $this->getSectionsText($this->content);
+        $current_font = $this->getDefaultFont($page);
 
-        foreach ($this->document->getObjects() as $obj) {
-            if ($obj instanceof Font) {
-                $current_font = $obj;
-                break;
-            }
-        }
+        $current_position_td = ['x' => false, 'y' => false];
+        $current_position_tm = ['x' => false, 'y' => false];
 
-        if ($current_font === null) {
-            $current_font = new Font($this->document);
-        }
-
-        $current_position_td = array('x' => false, 'y' => false);
-        $current_position_tm = array('x' => false, 'y' => false);
-
-        array_push(self::$recursionStack, $this->getUniqueId());
+        self::$recursionStack[] = $this->getUniqueId();
 
         foreach ($sections as $section) {
-
             $commands = $this->getCommandsText($section);
 
             foreach ($commands as $command) {
-
                 switch ($command[self::OPERATOR]) {
                     // set character spacing
                     case 'Tc':
@@ -287,55 +295,56 @@ class PDFObject
                     // move text current point
                     case 'Td':
                         $args = preg_split('/\s/s', $command[self::COMMAND]);
-                        $y    = array_pop($args);
-                        $x    = array_pop($args);
-                        if ((floatval($x) <= 0) ||
-                            ($current_position_td['y'] !== false && floatval($y) < floatval($current_position_td['y']))
+                        $y = array_pop($args);
+                        $x = array_pop($args);
+                        if (((float) $x <= 0) ||
+                            (false !== $current_position_td['y'] && (float) $y < (float) ($current_position_td['y']))
                         ) {
                             // vertical offset
                             $text .= "\n";
-                        } elseif ($current_position_td['x'] !== false && floatval($x) > floatval(
+                        } elseif (false !== $current_position_td['x'] && (float) $x > (float) (
                                 $current_position_td['x']
                             )
                         ) {
                             // horizontal offset
                             $text .= ' ';
                         }
-                        $current_position_td = array('x' => $x, 'y' => $y);
+                        $current_position_td = ['x' => $x, 'y' => $y];
                         break;
 
                     // move text current point and set leading
                     case 'TD':
                         $args = preg_split('/\s/s', $command[self::COMMAND]);
-                        $y    = array_pop($args);
-                        $x    = array_pop($args);
-                        if (floatval($y) < 0) {
+                        $y = array_pop($args);
+                        $x = array_pop($args);
+                        if ((float) $y < 0) {
                             $text .= "\n";
-                        } elseif (floatval($x) <= 0) {
+                        } elseif ((float) $x <= 0) {
                             $text .= ' ';
                         }
                         break;
 
                     case 'Tf':
-                        list($id,) = preg_split('/\s/s', $command[self::COMMAND]);
-                        $id           = trim($id, '/');
-                        if (!is_null($page)) {
-                            $current_font = $page->getFont($id);
+                        list($id) = preg_split('/\s/s', $command[self::COMMAND]);
+                        $id = trim($id, '/');
+                        if (null !== $page) {
+                            $new_font = $page->getFont($id);
+                            // If an invalid font ID is given, do not update the font.
+                            // This should theoretically never happen, as the PDF spec states for the Tf operator:
+                            // "The specified font value shall match a resource name in the Font entry of the default resource dictionary"
+                            // (https://www.adobe.com/content/dam/acom/en/devnet/pdf/pdfs/PDF32000_2008.pdf, page 435)
+                            // But we want to make sure that malformed PDFs do not simply crash.
+                            if (null !== $new_font) {
+                                $current_font = $new_font;
+                            }
                         }
                         break;
 
                     case "'":
                     case 'Tj':
-                        $command[self::COMMAND] = array($command);
+                        $command[self::COMMAND] = [$command];
+                        // no break
                     case 'TJ':
-                        // Skip if not previously defined, should never happened.
-                        if (is_null($current_font)) {
-                            // Fallback
-                            // TODO : Improve
-                            $text .= $command[self::COMMAND][0][self::COMMAND];
-                            break;
-                        }
-
                         $sub_text = $current_font->decodeText($command[self::COMMAND]);
                         $text .= $sub_text;
                         break;
@@ -347,21 +356,21 @@ class PDFObject
 
                     case 'Tm':
                         $args = preg_split('/\s/s', $command[self::COMMAND]);
-                        $y    = array_pop($args);
-                        $x    = array_pop($args);
-                        if ($current_position_tm['x'] !== false) {
-                            $delta = abs(floatval($x) - floatval($current_position_tm['x']));
+                        $y = array_pop($args);
+                        $x = array_pop($args);
+                        if (false !== $current_position_tm['x']) {
+                            $delta = abs((float) $x - (float) ($current_position_tm['x']));
                             if ($delta > 10) {
                                 $text .= "\t";
                             }
                         }
-                        if ($current_position_tm['y'] !== false) {
-                            $delta = abs(floatval($y) - floatval($current_position_tm['y']));
+                        if (false !== $current_position_tm['y']) {
+                            $delta = abs((float) $y - (float) ($current_position_tm['y']));
                             if ($delta > 10) {
                                 $text .= "\n";
                             }
                         }
-                        $current_position_tm = array('x' => $x, 'y' => $y);
+                        $current_position_tm = ['x' => $x, 'y' => $y];
                         break;
 
                     // set super/subscripting text rise
@@ -386,14 +395,13 @@ class PDFObject
                         break;
 
                     case 'Do':
-                        if (!is_null($page)) {
-                            $args    = preg_split('/\s/s', $command[self::COMMAND]);
-                            $id      = trim(array_pop($args), '/ ');
+                        if (null !== $page) {
+                            $args = preg_split('/\s/s', $command[self::COMMAND]);
+                            $id = trim(array_pop($args), '/ ');
                             $xobject = $page->getXObject($id);
 
-
-                             // @todo $xobject could be a ElementXRef object, which would then throw an error
-                             if ( is_object($xobject) && $xobject instanceof PDFObject && !in_array($xobject->getUniqueId(), self::$recursionStack) ) {
+                            // @todo $xobject could be a ElementXRef object, which would then throw an error
+                            if (\is_object($xobject) && $xobject instanceof self && !\in_array($xobject->getUniqueId(), self::$recursionStack)) {
                                 // Not a circular reference.
                                 $text .= $xobject->getText($page);
                             }
@@ -440,142 +448,135 @@ class PDFObject
         }
 
 
-        return $text . ' ';
+        return $text.' ';
     }
 
-	/**
-	 * @param Page
-	 *
-	 * @return array
-	 * @throws \Exception
-	 */
-	public function getTextArray(Page $page = null)
-	{
-		$text                = array();
-		$sections            = $this->getSectionsText($this->content);
-		$current_font        = new Font($this->document);
+    /**
+     * @param Page $page
+     *
+     * @return array
+     *
+     * @throws \Exception
+     */
+    public function getTextArray(Page $page = null)
+    {
+        $text = [];
+        $sections = $this->getSectionsText($this->content);
+        $current_font = new Font($this->document);
 
-		foreach ($sections as $section) {
+        foreach ($sections as $section) {
+            $commands = $this->getCommandsText($section);
 
-			$commands = $this->getCommandsText($section);
+            foreach ($commands as $command) {
+                switch ($command[self::OPERATOR]) {
+                    // set character spacing
+                    case 'Tc':
+                        break;
 
-			foreach ($commands as $command) {
+                    // move text current point
+                    case 'Td':
+                        break;
 
-				switch ($command[self::OPERATOR]) {
-					// set character spacing
-					case 'Tc':
-						break;
+                    // move text current point and set leading
+                    case 'TD':
+                        break;
 
-					// move text current point
-					case 'Td':
-						break;
+                    case 'Tf':
+                        if (null !== $page) {
+                            list($id) = preg_split('/\s/s', $command[self::COMMAND]);
+                            $id = trim($id, '/');
+                            $current_font = $page->getFont($id);
+                        }
+                        break;
 
-					// move text current point and set leading
-					case 'TD':
-						break;
+                    case "'":
+                    case 'Tj':
+                        $command[self::COMMAND] = [$command];
+                        // no break
+                    case 'TJ':
+                        $sub_text = $current_font->decodeText($command[self::COMMAND]);
+                        $text[] = $sub_text;
+                        break;
 
-					case 'Tf':
-						list($id,) = preg_split('/\s/s', $command[self::COMMAND]);
-						$id           = trim($id, '/');
-						$current_font = $page->getFont($id);
-						break;
+                    // set leading
+                    case 'TL':
+                        break;
 
-					case "'":
-					case 'Tj':
-						$command[self::COMMAND] = array($command);
-					case 'TJ':
-						// Skip if not previously defined, should never happened.
-						if (is_null($current_font)) {
-							// Fallback
-							// TODO : Improve
-							$text[] = $command[self::COMMAND][0][self::COMMAND];
-							break;
-						}
+                    case 'Tm':
+                        break;
 
-						$sub_text = $current_font->decodeText($command[self::COMMAND]);
-						$text[] = $sub_text;
-						break;
+                    // set super/subscripting text rise
+                    case 'Ts':
+                        break;
 
-					// set leading
-					case 'TL':
-						break;
+                    // set word spacing
+                    case 'Tw':
+                        break;
 
-					case 'Tm':
-						break;
+                    // set horizontal scaling
+                    case 'Tz':
+                        //$text .= "\n";
+                        break;
 
-					// set super/subscripting text rise
-					case 'Ts':
-						break;
+                    // move to start of next line
+                    case 'T*':
+                        //$text .= "\n";
+                        break;
 
-					// set word spacing
-					case 'Tw':
-						break;
+                    case 'Da':
+                        break;
 
-					// set horizontal scaling
-					case 'Tz':
-						//$text .= "\n";
-						break;
+                    case 'Do':
+                        if (null !== $page) {
+                            $args = preg_split('/\s/s', $command[self::COMMAND]);
+                            $id = trim(array_pop($args), '/ ');
+                            if ($xobject = $page->getXObject($id)) {
+                                $text[] = $xobject->getText($page);
+                            }
+                        }
+                        break;
 
-					// move to start of next line
-					case 'T*':
-						//$text .= "\n";
-						break;
+                    case 'rg':
+                    case 'RG':
+                        break;
 
-					case 'Da':
-						break;
+                    case 're':
+                        break;
 
-					case 'Do':
-						if (!is_null($page)) {
-							$args = preg_split('/\s/s', $command[self::COMMAND]);
-							$id   = trim(array_pop($args), '/ ');
-							if ($xobject = $page->getXObject($id)) {
-								$text[] = $xobject->getText($page);
-							}
-						}
-						break;
+                    case 'co':
+                        break;
 
-					case 'rg':
-					case 'RG':
-						break;
+                    case 'cs':
+                        break;
 
-					case 're':
-						break;
+                    case 'gs':
+                        break;
 
-					case 'co':
-						break;
+                    case 'en':
+                        break;
 
-					case 'cs':
-						break;
+                    case 'sc':
+                    case 'SC':
+                        break;
 
-					case 'gs':
-						break;
+                    case 'g':
+                    case 'G':
+                        break;
 
-					case 'en':
-						break;
+                    case 'V':
+                        break;
 
-					case 'sc':
-					case 'SC':
-						break;
+                    case 'vo':
+                    case 'Vo':
+                        break;
 
-					case 'g':
-					case 'G':
-						break;
+                    default:
+                }
+            }
+        }
 
-					case 'V':
-						break;
-
-					case 'vo':
-					case 'Vo':
-						break;
-
-					default:
-				}
-			}
-		}
-
-		return $text;
-	}
-
+        return $text;
+    }
 
     /**
      * @param string $text_part
@@ -585,15 +586,15 @@ class PDFObject
      */
     public function getCommandsText($text_part, &$offset = 0)
     {
-        $commands = $matches = array();
+        $commands = $matches = [];
 
-        while ($offset < strlen($text_part)) {
+        while ($offset < \strlen($text_part)) {
             $offset += strspn($text_part, "\x00\x09\x0a\x0c\x0d\x20", $offset);
             $char = $text_part[$offset];
 
             $operator = '';
-            $type     = '';
-            $command  = false;
+            $type = '';
+            $command = false;
 
             switch ($char) {
                 case '/':
@@ -605,8 +606,8 @@ class PDFObject
                     )
                     ) {
                         $operator = $matches[2];
-                        $command  = $matches[1];
-                        $offset += strlen($matches[0]);
+                        $command = $matches[1];
+                        $offset += \strlen($matches[0]);
                     } elseif (preg_match(
                         '/^\/([A-Z0-9\._,\+]+)\s+([A-Z]+)\s*/si',
                         substr($text_part, $offset),
@@ -614,8 +615,8 @@ class PDFObject
                     )
                     ) {
                         $operator = $matches[2];
-                        $command  = $matches[1];
-                        $offset += strlen($matches[0]);
+                        $command = $matches[1];
+                        $offset += \strlen($matches[0]);
                     }
                     break;
 
@@ -623,14 +624,14 @@ class PDFObject
                 case ']':
                     // array object
                     $type = $char;
-                    if ($char == '[') {
+                    if ('[' == $char) {
                         ++$offset;
                         // get elements
                         $command = $this->getCommandsText($text_part, $offset);
 
                         if (preg_match('/^\s*[A-Z]{1,2}\s*/si', substr($text_part, $offset), $matches)) {
                             $operator = trim($matches[0]);
-                            $offset += strlen($matches[0]);
+                            $offset += \strlen($matches[0]);
                         }
                     } else {
                         ++$offset;
@@ -643,24 +644,24 @@ class PDFObject
                     // array object
                     $type = $char;
                     ++$offset;
-                    if ($char == '<') {
-                        $strpos  = strpos($text_part, '>', $offset);
+                    if ('<' == $char) {
+                        $strpos = strpos($text_part, '>', $offset);
                         $command = substr($text_part, $offset, ($strpos - $offset));
-                        $offset  = $strpos + 1;
+                        $offset = $strpos + 1;
                     }
 
                     if (preg_match('/^\s*[A-Z]{1,2}\s*/si', substr($text_part, $offset), $matches)) {
                         $operator = trim($matches[0]);
-                        $offset += strlen($matches[0]);
+                        $offset += \strlen($matches[0]);
                     }
                     break;
 
                 case '(':
                 case ')':
                     ++$offset;
-                    $type   = $char;
+                    $type = $char;
                     $strpos = $offset;
-                    if ($char == '(') {
+                    if ('(' == $char) {
                         $open_bracket = 1;
                         while ($open_bracket > 0) {
                             if (!isset($text_part[$strpos])) {
@@ -669,37 +670,35 @@ class PDFObject
                             $ch = $text_part[$strpos];
                             switch ($ch) {
                                 case '\\':
-                                { // REVERSE SOLIDUS (5Ch) (Backslash)
+                                 // REVERSE SOLIDUS (5Ch) (Backslash)
                                     // skip next character
                                     ++$strpos;
                                     break;
-                                }
+
                                 case '(':
-                                { // LEFT PARENHESIS (28h)
+                                 // LEFT PARENHESIS (28h)
                                     ++$open_bracket;
                                     break;
-                                }
+
                                 case ')':
-                                { // RIGHT PARENTHESIS (29h)
+                                 // RIGHT PARENTHESIS (29h)
                                     --$open_bracket;
                                     break;
-                                }
                             }
                             ++$strpos;
                         }
                         $command = substr($text_part, $offset, ($strpos - $offset - 1));
-                        $offset  = $strpos;
+                        $offset = $strpos;
 
                         if (preg_match('/^\s*([A-Z\']{1,2})\s*/si', substr($text_part, $offset), $matches)) {
                             $operator = $matches[1];
-                            $offset += strlen($matches[0]);
+                            $offset += \strlen($matches[0]);
                         }
                     }
                     break;
 
                 default:
-
-                    if (substr($text_part, $offset, 2) == 'ET') {
+                    if ('ET' == substr($text_part, $offset, 2)) {
                         break;
                     } elseif (preg_match(
                         '/^\s*(?P<data>([0-9\.\-]+\s*?)+)\s+(?P<id>[A-Z]{1,3})\s*/si',
@@ -708,26 +707,26 @@ class PDFObject
                     )
                     ) {
                         $operator = trim($matches['id']);
-                        $command  = trim($matches['data']);
-                        $offset += strlen($matches[0]);
+                        $command = trim($matches['data']);
+                        $offset += \strlen($matches[0]);
                     } elseif (preg_match('/^\s*([0-9\.\-]+\s*?)+\s*/si', substr($text_part, $offset), $matches)) {
-                        $type    = 'n';
+                        $type = 'n';
                         $command = trim($matches[0]);
-                        $offset += strlen($matches[0]);
+                        $offset += \strlen($matches[0]);
                     } elseif (preg_match('/^\s*([A-Z\*]+)\s*/si', substr($text_part, $offset), $matches)) {
-                        $type     = '';
+                        $type = '';
                         $operator = $matches[1];
-                        $command  = '';
-                        $offset += strlen($matches[0]);
+                        $command = '';
+                        $offset += \strlen($matches[0]);
                     }
             }
 
-            if ($command !== false) {
-                $commands[] = array(
-                    self::TYPE     => $type,
+            if (false !== $command) {
+                $commands[] = [
+                    self::TYPE => $type,
                     self::OPERATOR => $operator,
-                    self::COMMAND  => $command,
-                );
+                    self::COMMAND => $command,
+                ];
             } else {
                 break;
             }
@@ -737,49 +736,49 @@ class PDFObject
     }
 
     /**
-     * @param $document Document
-     * @param $header   Header
-     * @param $content  string
+     * @param string $content
      *
      * @return PDFObject
      */
-    public static function factory(Document $document, Header $header, $content)
-    {
+    public static function factory(
+        Document $document,
+        Header $header,
+        $content,
+        Config $config = null
+    ) {
         switch ($header->get('Type')->getContent()) {
             case 'XObject':
                 switch ($header->get('Subtype')->getContent()) {
                     case 'Image':
-                        return new Image($document, $header, $content);
+                        return new Image($document, $header, $content, $config);
 
                     case 'Form':
-                        return new Form($document, $header, $content);
-
-                    default:
-                        return new PDFObject($document, $header, $content);
+                        return new Form($document, $header, $content, $config);
                 }
-                break;
+
+                return new self($document, $header, $content, $config);
 
             case 'Pages':
-                return new Pages($document, $header, $content);
+                return new Pages($document, $header, $content, $config);
 
             case 'Page':
-                return new Page($document, $header, $content);
+                return new Page($document, $header, $content, $config);
 
             case 'Encoding':
-                return new Encoding($document, $header, $content);
+                return new Encoding($document, $header, $content, $config);
 
             case 'Font':
-                $subtype   = $header->get('Subtype')->getContent();
-                $classname = '\Smalot\PdfParser\Font\Font' . $subtype;
+                $subtype = $header->get('Subtype')->getContent();
+                $classname = '\Smalot\PdfParser\Font\Font'.$subtype;
 
                 if (class_exists($classname)) {
-                    return new $classname($document, $header, $content);
-                } else {
-                    return new Font($document, $header, $content);
+                    return new $classname($document, $header, $content, $config);
                 }
 
+                return new Font($document, $header, $content, $config);
+
             default:
-                return new PDFObject($document, $header, $content);
+                return new self($document, $header, $content, $config);
         }
     }
 
