@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,7 +26,6 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser\Element;
@@ -34,8 +34,6 @@ use Smalot\PdfParser\Document;
 
 /**
  * Class ElementHexa
- *
- * @package Smalot\PdfParser\Element
  */
 class ElementHexa extends ElementString
 {
@@ -44,15 +42,15 @@ class ElementHexa extends ElementString
      * @param Document $document
      * @param int      $offset
      *
-     * @return bool|ElementHexa
+     * @return bool|ElementHexa|ElementDate
      */
     public static function parse($content, Document $document = null, &$offset = 0)
     {
         if (preg_match('/^\s*\<(?P<name>[A-F0-9]+)\>/is', $content, $match)) {
-            $name    = $match['name'];
-            $offset += strpos($content, '<' . $name) + strlen($name) + 2; // 1 for '>'
+            $name = $match['name'];
+            $offset += strpos($content, '<'.$name) + \strlen($name) + 2; // 1 for '>'
             // repackage string as standard
-            $name    = '(' . self::decode($name, $document) . ')';
+            $name = '('.self::decode($name, $document).')';
             $element = false;
 
             if (!($element = ElementDate::parse($name, $document))) {
@@ -71,18 +69,18 @@ class ElementHexa extends ElementString
      */
     public static function decode($value, Document $document = null)
     {
-        $text   = '';
-        $length = strlen($value);
+        $text = '';
+        $length = \strlen($value);
 
-        if (substr($value, 0, 2) == '00') {
+        if ('00' === substr($value, 0, 2)) {
             for ($i = 0; $i < $length; $i += 4) {
                 $hex = substr($value, $i, 4);
-                $text .= '&#' . str_pad(hexdec($hex), 4, '0', STR_PAD_LEFT) . ';';
+                $text .= '&#'.str_pad(hexdec($hex), 4, '0', STR_PAD_LEFT).';';
             }
         } else {
             for ($i = 0; $i < $length; $i += 2) {
                 $hex = substr($value, $i, 2);
-                $text .= chr(hexdec($hex));
+                $text .= \chr(hexdec($hex));
             }
         }
 
