@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,31 +26,23 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser\Element;
 
 use Smalot\PdfParser\Element;
-use Smalot\PdfParser\Document;
 
 /**
  * Class ElementMissing
  */
 class ElementMissing extends Element
 {
-    /**
-     * @param string   $value
-     * @param Document $document
-     */
-    public function __construct($value, Document $document = null)
+    public function __construct()
     {
         parent::__construct(null, null);
     }
 
     /**
-     * @param mixed $value
-     *
      * @return bool
      */
     public function equals($value)
@@ -58,8 +51,6 @@ class ElementMissing extends Element
     }
 
     /**
-     * @param mixed $value
-     *
      * @return bool
      */
     public function contains($value)
