@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,34 +26,28 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser\Element;
 
-use Smalot\PdfParser\Element;
 use Smalot\PdfParser\Document;
+use Smalot\PdfParser\Element;
 use Smalot\PdfParser\Font;
 
 /**
  * Class ElementName
- *
- * @package Smalot\PdfParser\Element
  */
 class ElementName extends Element
 {
     /**
-     * @param string   $value
-     * @param Document $document
+     * @param string $value
      */
-    public function __construct($value, Document $document = null)
+    public function __construct($value)
     {
         parent::__construct($value, null);
     }
 
     /**
-     * @param mixed $value
-     *
      * @return bool
      */
     public function equals($value)
@@ -69,12 +64,12 @@ class ElementName extends Element
      */
     public static function parse($content, Document $document = null, &$offset = 0)
     {
-        if (preg_match('/^\s*\/(?P<name>[A-Z0-9\-\+,#\.]+)/is', $content, $match)) {
-            $name   = $match['name'];
-            $offset += strpos($content, $name) + strlen($name);
-            $name   = Font::decodeEntities($name);
+        if (preg_match('/^\s*\/([A-Z0-9\-\+,#\.]+)/is', $content, $match)) {
+            $name = $match[1];
+            $offset += strpos($content, $name) + \strlen($name);
+            $name = Font::decodeEntities($name);
 
-            return new self($name, $document);
+            return new self($name);
         }
 
         return false;
