@@ -6,6 +6,7 @@
  *
  * @author  Sébastien MALOT <sebastien@malot.fr>
  * @date    2017-01-03
+ *
  * @license LGPLv3
  * @url     <https://github.com/smalot/pdfparser>
  *
@@ -25,46 +26,44 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.
  *  If not, see <http://www.pdfparser.org/sites/default/LICENSE.txt>.
- *
  */
 
 namespace Smalot\PdfParser;
 
 /**
  * Class Pages
- *
- * @package Smalot\PdfParser
  */
 class Pages extends PDFObject
 {
     /**
      * @param bool $deep
      *
+     * @todo Objects other than Pages or Page might need to be treated specifically in order to get Page objects out of them,
+     * see https://github.com/smalot/pdfparser/issues/331
+     *
      * @return array
      */
     public function getPages($deep = false)
     {
-        if ($this->has('Kids')) {
+        if (!$this->has('Kids')) {
+            return [];
+        }
 
-            if (!$deep) {
-                return $this->get('Kids')->getContent();
-            } else {
-                $kids  = $this->get('Kids')->getContent();
-                $pages = array();
+        if (!$deep) {
+            return $this->get('Kids')->getContent();
+        }
 
-                foreach ($kids as $kid) {
+        $kids = $this->get('Kids')->getContent();
+        $pages = [];
 
-                    if ($kid instanceof Pages) {
-                        $pages = array_merge($pages, $kid->getPages(true));
-                    } else {
-                        $pages[] = $kid;
-                    }
-                }
-
-                return $pages;
+        foreach ($kids as $kid) {
+            if ($kid instanceof self) {
+                $pages = array_merge($pages, $kid->getPages(true));
+            } elseif ($kid instanceof Page) {
+                $pages[] = $kid;
             }
         }
 
-        return array();
+        return $pages;
     }
 }
